requirements.txt
flask
python-dotenv
google-genai
requests
gunicorn
psycopg2-binary